Rename context to vmcontext
